Peanut/requirements.txt
streamlit
plotly
kaleido
pandasai
pandasai[langchain]
langchain_groq
pygwalker